move implementation to .cpp file on "[ExecuTorch] generalize tests for unary_ufunc_realhb_to_floath ops (1/2)"
We have 20-odd ops that just map a unary function over each
element. We can share the test infrastructure rather than create a
bunch of similar-looking tests. Note that I attempted to minimize the
amount of code inside `IMPLEMENT_UNARY_UFUNC_REALHB_TO_FLOATH_TEST`;
it is just a bunch of `TEST_F` calls that call into plain old C++
functions, so we haven't merely copied the tests into a macro.

This diff only makes one op adopt the test to prove it works; following diff(s) will roll out fully.

Differential Revision: [D63431290](https://our.internmc.facebook.com/intern/diff/D63431290/)

[ghstack-poisoned]
